fix: renamed the python file of global_element_ids.py
